refactor(runtime): share run-fact-to-progress-event subscription helper
LegacyProgressEventProjection.ts and ProgressBackend.ts each hand-rolled a
byte-identical SessionEventHub subscription (same run-scope filter, same
five fact types, same projectRunFactToProgressEvent call), differing only
in the sink. Extract subscribeRunFactsAsProgressEvents() into
sessionProgressEventProjection.ts so the filter and projection can't
silently diverge between the two hosts. No behavior change.
