+++ src/agent/runtime/sessionProgressEventProjection.ts
@@ -12,7 +12,7 @@ import {
 import { isObject } from '@utils/core';
 
 import { fromRunFactDomainKey } from './runFactEvents';
-import type { SessionFact } from './SessionEventHub';
+import type { SessionEventHub, SessionFact } from './SessionEventHub';
 
 export type ProjectedProgressEvent = {
   [K in ProgressEvent]: {
@@ -152,4 +152,35 @@ export function projectRunFactToProgressEvent(
   }
 
   return undefined;
+}
+
+const RUN_FACT_PROGRESS_EVENT_TYPES: readonly AgentEvent['type'][] = [
+  'domain',
+  'usage',
+  'stage.start',
+  'child.activity',
+  'process.output',
+];
+
+/**
+ * Subscribe to the run-scoped facts that project onto legacy progress
+ * events, forwarding each successfully projected event to `onProjected`.
+ * Shared by `LegacyProgressEventProjection` and `ProgressBackend` so their
+ * run-fact filter and projection can't silently diverge.
+ */
+export function subscribeRunFactsAsProgressEvents(
+  events: SessionEventHub,
+  onProjected: (projected: ProjectedProgressEvent) => void,
+): () => void {
+  return events.subscribe(
+    (sessionEvent) => {
+      if (sessionEvent.scope !== 'run') return;
+      const projected = projectRunFactToProgressEvent(
+        sessionEvent.streamId,
+        sessionEvent.event,
+      );
+      if (projected) onProjected(projected);
+    },
+    { scope: 'run', types: RUN_FACT_PROGRESS_EVENT_TYPES },
+  );
 }
